Task investigation summary component

+++ FloodOnlineReportingTool.Public/Components/Pages/FloodReport/Investigation/Summary.razor.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Components.Forms;
 using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
+using Microsoft.Identity.Web;
 
 namespace FloodOnlineReportingTool.Public.Components.Pages.FloodReport.Investigation;
 
@@ -49,6 +50,8 @@ public partial class Summary(
     private IReadOnlyCollection<RecordStatus> _recordStatuses = [];
     private IReadOnlyCollection<FloodMitigation> _floodMitigations = [];
 
+    private InvestigationDto? _investigationDto;
+
     protected override void OnInitialized()
     {
         // Setup model and edit context
@@ -77,24 +80,26 @@ public partial class Summary(
     {
         if (firstRender)
         {
-            var userId = await AuthenticationState.IdentityUserId() ?? Guid.Empty;
-            var eligibilityCheck = await eligibilityCheckRepository.ReportedByUser(userId, _cts.Token);
-            var isInternal = eligibilityCheck?.IsInternal() == true;
+            bool isInternal = false;
+            var userId = await GetUserIdAsGuid();
+            if (userId.HasValue)
+            {
+                var eligibilityCheck = await eligibilityCheckRepository.ReportedByUser(userId.Value, _cts.Token);
+                isInternal = eligibilityCheck?.IsInternal() == true;
+            }
 
             _floodProblems = await GetInvestigationFloodProblems();
             _recordStatuses = await GetInvestigationRecordStatuses();
             _floodMitigations = await GetInvestigationFloodMitigations();
 
-            var investigation = await GetInvestigation();
+            _investigationDto = await GetInvestigation();
 
-            await CreateSummary(investigation, isInternal);
+            await CreateSummary(_investigationDto, isInternal);
 
             _editContext.Validate();
             _isLoading = false;
             StateHasChanged();
             _editContext.NotifyValidationStateChanged();
-
-            
         }
     }
 
@@ -206,12 +211,28 @@ public partial class Summary(
 
     private async Task SaveInvestigation()
     {
+        if (_investigationDto is null)
+        {
+            logger.LogError("Investigation information was not found. Investigation: {Investigation}", _investigationDto);
+        }
+
+        var userId = await GetUserIdAsGuid();
+        if (userId is null)
+        {
+            logger.LogError("User ID was not found.");
+        }
+
+        if (_investigationDto is null || userId is null)
+        {
+            _messageStore.Add(errorField, "There was a problem saving the investigation. Please try again but if this issue happens again then please report a bug.");
+            _editContext.NotifyValidationStateChanged();
+            return;
+        }
+
         logger.LogDebug("Saving investigation information..");
         try
         {
-            var userId = await AuthenticationState.IdentityUserId();
-            var investigation = await GetInvestigation();
-            await investigationRepository.CreateForUser(userId.Value, investigation, _cts.Token);
+            await investigationRepository.CreateForUser(userId.Value, _investigationDto, _cts.Token);
 
             // Clear the session data
             await protectedSessionStorage.DeleteAsync(SessionConstants.Investigation);
@@ -227,30 +248,12 @@ public partial class Summary(
         }
     }
 
-    private async Task<EligibilityCheckDto> GetEligibilityCheck()
-    {
-        var data = await protectedSessionStorage.GetAsync<EligibilityCheckDto>(SessionConstants.EligibilityCheck);
-        if (data.Success)
-        {
-            if (data.Value != null)
-            {
-                return data.Value;
-            }
-        }
-
-        logger.LogDebug("Eligibility Check was not found in the protected storage.");
-        return new();
-    }
-
     private async Task<InvestigationDto> GetInvestigation()
     {
         var data = await protectedSessionStorage.GetAsync<InvestigationDto>(SessionConstants.Investigation);
-        if (data.Success)
+        if (data.Success && data.Value is not null)
         {
-            if (data.Value != null)
-            {
-                return data.Value;
-            }
+            return data.Value;
         }
 
         logger.LogWarning("Investigation was not found in the protected storage.");
@@ -378,5 +381,20 @@ public partial class Summary(
             .Where(o => ids.Contains(o.Id))
             .Select(o => o.TypeName ?? ""),
         ];
+    }
+
+    private async Task<string?> GetUserId()
+    {
+        if (AuthenticationState is null)
+        {
+            return null;
+        }
+        var authState = await AuthenticationState;
+        return authState.User.GetObjectId();
+    }
+
+    private async Task<Guid?> GetUserIdAsGuid()
+    {
+        return Guid.TryParse(await GetUserId(), out var userId) ? userId : null;
     }
 }
